+++ modules/reporting/bingraph.py
@@ -62,7 +62,7 @@ class BinGraph(Report):
             except Exception as e:
                 log.info(e)
 
-            for key in ("dropped", "procdump", "CAPE"):
+            for key in ("dropped", "procdump"):
                 for block in results.get(key, []) or []:
                     if (
                         block.get("size", 0) != 0
@@ -76,9 +76,26 @@ class BinGraph(Report):
                             path = block["path"]
                         if not path:
                             continue
-                        bingraph_args_dict.update(
-                            {"prefix": block["sha256"], "files": [path], "save_dir": bingraph_path,}
-                        )
+                        bingraph_args_dict.update({"prefix": block["sha256"], "files": [path], "save_dir": bingraph_path})
+                        try:
+                            bingraph_gen(bingraph_args_dict)
+                        except Exception as e:
+                            log.warning("Can't generate report for {}: {}".format(path, e))
+
+                for block in results.get("CAPE", {}).get("payloads") or []:
+                    if (
+                        block.get("size", 0) != 0
+                        and block.get("type", "") not in excluded_filetypes
+                        and not os.path.exists(os.path.join(bingraph_path, "{}-ent.svg".format(block["sha256"])))
+                    ):
+                        path = ""
+                        if block.get("file", False):
+                            path = block["file"]
+                        elif block.get("path", False):
+                            path = block["path"]
+                        if not path:
+                            continue
+                        bingraph_args_dict.update({"prefix": block["sha256"], "files": [path], "save_dir": bingraph_path})
                         try:
                             bingraph_gen(bingraph_args_dict)
                         except Exception as e: