+++ web/api/views.py
@@ -560,7 +560,7 @@ def tasks_vtdl(request):
         if opts:
             opt_apikey = opts.get("apikey", False)
 
-        if (not settings.VTDL_PRIV_KEY and not settings.VTDL_INTEL_KEY) or not settings.VTDL_PATH or not opt_apikey:
+        if settings.VTDL_KEY or not settings.VTDL_PATH or not opt_apikey:
             resp = {"error": True, "error_value": "You specified VirusTotal but must edit the file and specify your VTDL_KEY variable and VTDL_PATH base directory"}
             return jsonize(resp, response=True)
 
@@ -1942,11 +1942,18 @@ def tasks_config(request, task_id, cape_name=False):
                 # In case compress results processing module is not enabled
                 pass
             data = []
-            for cape in buf["CAPE"]:
-                if isinstance(cape, dict) and cape.get("cape_config"):
-                    if cape_name and cape.get("cape_name", "") == cape_name:
-                        return jsonize(cape["cape_config"], response=True)
-                    data.append(cape)
+
+            if not isinstance(buf["CAPE"], list) and buf["CAPE"].get("cape_configs"):
+                if cape_name and buf["CAPE"]["cape_configs"].get("cape_name", "") == cape_name:
+                    return jsonize({cape_name.lower(): buf["CAPE"]["cape_configs"][cape_name]}, response=True)
+                data = buf["CAPE"]["cape_configs"]
+            # ToDo remove in v3
+            elif buf["CAPE"]:
+                for cape in buf["CAPE"]:
+                    if isinstance(cape, dict) and cape.get("cape_config"):
+                        if cape_name and cape.get("cape_name", "") == cape_name:
+                            return jsonize(cape["cape_config"], response=True)
+                        data.append(cape)
             if data:
                 resp = {"error": False, "configs": data}
             else: